PS/2 to MSX keyboard adapter with USB Console
USB is implemented to be used as console, with the target to allow to upload the database through USB.
USB is available as DFU (Device Softwqre Update), so you will only need a USB C cable => the same cable used in nowadays mobile phones.
This is a bare metal design adapter based on Blue Pill or Black Pill F401Mini to attach a PS/2 keyboard to a MSX [or other computers that use a up to 15 (input) x 8 (reading) matrix] with 4 bit BCD input. It has expandable capabilities, using LIBOPENCM3 library.
This unit has a highly configurable capability, as it has an user updateable database, without need to recompile and upload a new firmwware, to translate from PC keyboard to MSX.
Now with doxygen ready in the source files.
Indentation to spaces corrected.