Database Compiler Excel File
Database Compiler Excel File update.
This file has three sheets: HotBit, Expert and Your MSX Definitions. All three are available to compile the Database, in Generated table sheet. You may use HotBit Deinitions as a guide, since it is correct and it is what I use.
After fill data, click on the MSX keyboard picture to run the excel VBA macro, so you have to enable macros in your excel config.
This file is also able to run in Open Office / LibreOffice family, as compatibilty mode. It will take longer time to run, because it runs on VBA compatibilty mode.